LayoutTests/imported/w3c/web-platform-tests/scroll-animations/view-timelines/view-timeline-subject-size-changes-expected.txt


FAIL View timeline with subject size change after the creation of the animation assert_equals: Effect at the start of the active phase expected "0.3" but got "0.609278"
